+++ lib/rails_tracepoint_stack/configuration.rb
@@ -1,9 +1,10 @@
 module RailsTracepointStack
   class Configuration
-    attr_accessor :ignore_patterns
+    attr_accessor :ignore_patterns, :logger
 
     def initialize
       @ignore_patterns = []
+      @logger = nil
     end
   end
 end