+++ changelog.md
@@ -1,5 +1,14 @@
 version
 
+<b> 0.2.0 <b>
+
+changes:
+
+- Add configuration of custom log patterns to be ignored
+- Add Rspec and Rake
+
+version
+
 <b> 0.1.4 <b>
 
 changes: